+++ ledger/ledger-api-tests/suites/src/main/scala/com/daml/ledger/api/testtool/suites/v1_8/PartyManagementServiceIT.scala
@@ -6,7 +6,6 @@ package com.daml.ledger.api.testtool.suites.v1_8
 import com.daml.error.definitions.LedgerApiErrors
 import com.daml.ledger.api.testtool.infrastructure.Allocation._
 import com.daml.ledger.api.testtool.infrastructure.Assertions._
-import com.daml.ledger.api.testtool.infrastructure.LedgerTestSuite
 import com.daml.ledger.api.v1.admin.party_management_service.{
   AllocatePartyRequest,
   AllocatePartyResponse,
@@ -21,17 +20,12 @@ import scalaz.Tag
 import scalaz.syntax.tag.ToTagOps
 import java.util.regex.Pattern
 
-import com.daml.ledger.api.testtool.suites.v1_8.objectmeta.PartyManagementObjectMetaTests
 import com.daml.ledger.api.v1.admin.object_meta.ObjectMeta
 import com.daml.ledger.client.binding.Primitive
 
 import scala.util.Random
 
-final class PartyManagementServiceIT
-    extends LedgerTestSuite
-    with PartyManagementItUtils
-    with PartyManagementServiceUpdateRpcTests
-    with PartyManagementObjectMetaTests {
+final class PartyManagementServiceIT extends PartyManagementITBase {
 
   test(
     "PMNonEmptyParticipantID",
